Move Passenger57 class to a separate file
