+++ src/roulette.py
@@ -4,52 +4,6 @@ from dataclasses import dataclass
 from abc import ABC, abstractmethod
 
 
-class Passenger57(Player):
-    """
-    :class:`Passenger57` constructs a :class:`Bet` instance based on the :class:`Outcome` object
-    named :samp:`"Black"`. This is a very persistent player.
-
-    .. attribute:: black
-
-       This is the outcome on which this player focuses their betting.
-
-       This :class:`Player` will get this from the :class:`Wheel` using a well-known bet name.
-
-    .. attribute:: table
-
-       The :class:`Table` that is used to place individual :class:`Bet` instances.
-
-    """
-
-    def __init__(self, table: Table, wheel: Wheel) -> None:
-        """
-        Constructs the :class:`Player` instance with a specific table for placing bets. This also
-        creates the “black” :class:`Outcome`. This is saved in a variable named
-        **Passenger57.black** for use in creating bets.
-
-        :param table: The :class:`Table` instance on which bets are placed.
-        :param wheel: The :class:`Wheel` instance which defines all :class:`Outcome` instances.
-        """
-
-        super().__init__(table)
-        self.table = table
-        self.wheel = wheel
-        self.black = self.wheel.getOutcome("Black")
-
-    def placeBets(self) -> None:
-        """
-        Updates the :class:`Table` object with the various bets. This version creates a :class:`Bet`
-        instance from the “Black” :class:`Outcome` instance. It uses **Table.placeBet()** to
-        place that bet.
-
-        """
-
-        bet_amount = 20
-        bet = Bet(bet_amount, self.black)
-        self.table.placeBet(bet)
-        self.stake -= bet_amount
-
-
 def main() -> None:  # pragma: no cover
     """
     A main application function that creates the necessary objects, runs the Simulator’s gather()
